% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST.R
\name{activateST}
\alias{activateST}
\title{Activate st-storage in an Antares study}
\usage{
activateST(opts = antaresRead::simOptions(), quietly = !interactive())
}
\arguments{
\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{quietly}{Display or not a message to the user if success.}
}
\value{
An updated list containing various information about the simulation.
}
\description{
Activate st-storage in an Antares study
}
\examples{
\dontrun{

library(antaresEditObject)
tmp <- tempfile()
createStudy(path = tmp)
opts <- antaresRead::setSimulationPath(tmp)
activateST()

# then you can use createClusterST()...

}
}
