% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeHydroValues.R
\name{add_week_number_column_to_ts}
\alias{add_week_number_column_to_ts}
\title{Add week number column to a data.time of time series type}
\usage{
add_week_number_column_to_ts(xts)
}
\arguments{
\item{xts}{a data.table of time series type.}
}
\value{
the data.table xts with a new column week.
}
\description{
If timeId column exists, add a week number column. A week is 168 consecutive hours (= 24 * 7).
}
