% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createArea.R
\name{adequacyOptions}
\alias{adequacyOptions}
\title{Adequacy patch parameters for creating an area}
\usage{
adequacyOptions(adequacy_patch_mode = "outside")
}
\arguments{
\item{adequacy_patch_mode}{character, default to "outside"}
}
\value{
a named list
}
\description{
Adequacy patch parameters for creating an area
}
\examples{
adequacyOptions()
}
