% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antaresEditObject-package.R
\docType{package}
\name{antaresEditObject-package}
\alias{antaresEditObject}
\alias{antaresEditObject-package}
\title{antaresEditObject: Edit an 'Antares' Simulation}
\description{
Edit an 'Antares' simulation before running it : create new areas, links, thermal clusters or binding constraints or edit existing ones. Update 'Antares' general & optimization settings. 'Antares' is an open source power system generator, more information available here : \url{https://antares-simulator.org/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rte-antares-rpackage/antaresEditObject}
  \item \url{https://rte-antares-rpackage.github.io/antaresEditObject/}
  \item Report bugs at \url{https://github.com/rte-antares-rpackage/antaresEditObject/issues}
}

}
\author{
\strong{Maintainer}: Tatiana Vargas \email{tatiana.vargas@rte-france.com}

Authors:
\itemize{
  \item Frederic Breant
  \item Victor Perrier
}

Other contributors:
\itemize{
  \item Baptiste Seguinot [contributor]
  \item Benoit Thieurmel [contributor]
  \item Titouan Robert [contributor]
  \item Jalal-Edine Zawam [contributor]
  \item Etienne Sanchez [contributor]
  \item Janus De Bondt [contributor]
  \item Assil Mansouri [contributor]
  \item Abdallah Mahoudi [contributor]
  \item Clement Berthet [contributor]
  \item Kamel Kemiha [contributor]
  \item Nicolas Boitard [contributor]
  \item RTE [copyright holder]
}

}
\keyword{internal}
