% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIni.R
\name{antaresRead-reexports}
\alias{antaresRead-reexports}
\alias{readIni}
\alias{readIniFile}
\alias{readIniAPI}
\title{Re-exports from antaresRead}
\usage{
readIni(pathIni, opts = antaresRead::simOptions(), default_ext = ".ini")

readIniFile(file, stringsAsFactors = FALSE)

readIniAPI(study_id, path, host, token = NULL)
}
\description{
Re-exports from antaresRead
}
\keyword{internal}
