% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupSimulation.R
\name{backupSimulation}
\alias{backupSimulation}
\title{Create a .tar.gz file with an Antares Study}
\usage{
backupSimulation(backupfile, opts = antaresRead::simOptions())
}
\arguments{
\item{backupfile}{Name of the backup, without extension.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
TRUE or FALSE accordingly to the success or not of the backup
}
\description{
Create a .tar.gz file with an Antares Study
}
\examples{
\dontrun{
backupStudy()
}
}
