% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeCluster.R
\name{check_cluster_in_binding_constraint}
\alias{check_cluster_in_binding_constraint}
\title{Check if a cluster is referenced in a binding constraint as a coefficient.}
\usage{
check_cluster_in_binding_constraint(
  area,
  cluster_name,
  add_prefix,
  opts = antaresRead::simOptions()
)
}
\arguments{
\item{area}{The area where to create the cluster.}

\item{cluster_name}{Name for the cluster, it will prefixed by area name, unless you set \code{add_prefix = FALSE}.}

\item{add_prefix}{If \code{TRUE} (the default), \code{cluster_name} will be prefixed by area name.}

\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}}
}
\value{
An updated list containing various information about the simulation.
}
\description{
Check if a cluster is referenced in a binding constraint as a coefficient.
}
