% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeHydroValues.R
\name{check_consistency_reservoir_values}
\alias{check_consistency_reservoir_values}
\title{For a given area, check consistency between reservoir and reservoir capacity values}
\usage{
check_consistency_reservoir_values(area, new_data, prev_data)
}
\arguments{
\item{area}{The area where to run the check.}

\item{new_data}{The new list of parameters.}

\item{prev_data}{The previous data found in hydro.ini.}
}
\description{
For a given area, check consistency between reservoir and reservoir capacity values
}
