% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeLink.R
\name{check_link_in_binding_constraint}
\alias{check_link_in_binding_constraint}
\title{Check if a link is referenced in a binding constraint as a coefficient.}
\usage{
check_link_in_binding_constraint(from, to, opts = antaresRead::simOptions())
}
\arguments{
\item{from, to}{The two areas linked together.}

\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}}
}
\value{
An updated list containing various information about the simulation.
}
\description{
Check if a link is referenced in a binding constraint as a coefficient.
}
