% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeHydroValues.R
\name{check_mingen_vs_hydro_storage}
\alias{check_mingen_vs_hydro_storage}
\title{Check if mingen data and maxpower data are consistent.}
\usage{
check_mingen_vs_hydro_storage(area, opts = antaresRead::simOptions())
}
\arguments{
\item{area}{The area to check.}

\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}.}
}
\description{
At each hourly time step, mingen must be less or equal than maxpower.
}
