% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeHydroValues.R
\name{check_mingen_vs_maxpower}
\alias{check_mingen_vs_maxpower}
\title{Check if mingen data and maxpower data are consistent}
\usage{
check_mingen_vs_maxpower(area, opts = antaresRead::simOptions())
}
\arguments{
\item{area}{The area where to check the data.}

\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}.}
}
\value{
a list containing the boolean if the check is ok and the message to display.
}
\description{
At each hourly time step, mingen must be less or equal than generatingMaxPower.
}
\note{
Function called only for an \strong{Antares version >= 860}.
}
