% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeArea.R
\name{checkRemoveArea}
\alias{checkRemoveArea}
\title{Seek for a removed area}
\usage{
checkRemoveArea(area, opts = antaresRead::simOptions())
}
\arguments{
\item{area}{An area}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
a named list with two elements
}
\description{
Check if it remains trace of a deleted area in the input folder
}
\examples{
\dontrun{
checkRemoveArea("myarea")
}
}
