% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanUpOutput.R
\name{cleanUpOutput}
\alias{cleanUpOutput}
\title{Clean up output based on geographic trimming}
\usage{
cleanUpOutput(areas = NULL, opts = simOptions())
}
\arguments{
\item{areas}{Character. vector of areas (folders). Links will also be cleaned based on getLinks() results}

\item{opts}{List. simulation options}
}
\description{
Clean up output based on geographic trimming
}
