% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTimeStepFromHourly.R
\name{computeOtherFromHourlyMulti}
\alias{computeOtherFromHourlyMulti}
\title{Compute daily, weekly, monthly and annual mc-ind from hourly data multiyear. (new)}
\usage{
computeOtherFromHourlyMulti(
  opts = simOptions(),
  areas = "all",
  type = c("areas", "links", "clusters"),
  timeStep = c("daily", "monthly", "annual", "weekly"),
  mcYears = simOptions()$mcYears,
  writeOutput = FALSE,
  nbcl = 8,
  verbose = FALSE
)
}
\arguments{
\item{opts}{study opts}

\item{areas}{vector of areas}

\item{type}{type of aggregation}

\item{timeStep}{timestep of aggregation (daily, monthly and annual, NO weekly)}

\item{mcYears}{vector of years to compute}

\item{writeOutput}{boolean to write data in mc-ind folder}

\item{nbcl}{number of cpu cores for parallelization}

\item{verbose}{logical for printing output}
}
\description{
Compute daily, weekly, monthly and annual mc-ind from hourly data multiyear. (new)
}
\note{
Recommended only with studies spanning on two years.
}
\seealso{
\code{\link{computeOtherFromHourlyYear}}
}
