% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTimeStepFromHourly.R
\name{computeOtherFromHourlyYear}
\alias{computeOtherFromHourlyYear}
\title{Compute daily, weekly, monthly and annual mc-ind from hourly data for one year. (new)}
\usage{
computeOtherFromHourlyYear(
  mcYear,
  type,
  areas = "all",
  opts = simOptions(),
  timeStep = c("daily", "monthly", "annual", "weekly"),
  writeOutput = F
)
}
\arguments{
\item{mcYear}{vector of years to compute}

\item{type}{type of data (areas, links, clusters, clustersRes)}

\item{areas}{vector of areas. links type will use getLinks() to get data.}

\item{opts}{study opts}

\item{timeStep}{timestep of aggregation (daily, monthly and annual, NO weekly)}

\item{writeOutput}{boolean to write data in mc-ind folder}
}
\description{
Compute daily, weekly, monthly and annual mc-ind from hourly data for one year. (new)
}
\note{
Recommended only with studies spanning on two years.
}
\seealso{
\code{\link{computeOtherFromHourlyMulti}}
}
