% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTimeStampFromHourly.R
\name{computeTimeStampFromHourly}
\alias{computeTimeStampFromHourly}
\title{Compute daily, weekly, monthly and annual mc-ind from hourly data.}
\usage{
computeTimeStampFromHourly(
  opts,
  mcYears = "all",
  nbcl = 8,
  verbose = 1,
  type = c("areas", "links", "clusters")
)
}
\arguments{
\item{opts}{opts simulation path.}

\item{mcYears}{mcYears to compute.}

\item{nbcl}{number of thread for parallel computing.}

\item{verbose}{verbose for execution.}

\item{type}{type of file to compute.}
}
\description{
\ifelse{html}{\figure{badge_api_no.svg}{options: alt='Antares API NO'}}{Antares API: \strong{NO}}

Compute daily, weekly, monthly and annual mc-ind from hourly data.
}
\note{
Deprecated on studies v8 or higher.
}
\examples{
\dontrun{

library(antaresEditObject)
opts <- setSimulationPath("my_study")
computeTimeStampFromHourly(opts)

}

}
