% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateAdequacySettings.R
\name{convertConfigToAdq}
\alias{convertConfigToAdq}
\title{Read adequacy patch config.yml into Antares (v8.5+)}
\usage{
convertConfigToAdq(opts = simOptions(), path = "default")
}
\arguments{
\item{opts}{List. study options.}

\item{path}{Character. path to config.yml. Default points to "/user/adequacypatch/" in study}
}
\description{
Use this function to load config.yml used in older Antares versions for adequacy patch.
Areas in config will be updated to be included in adequacy patch perimeter.
}
\seealso{
\code{\link{updateAdequacySettings}}
}
