% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyOutput.R
\name{copyOputput}
\alias{copyOputput}
\title{Create An Area In An Antares Study}
\usage{
copyOputput(opts, extname)
}
\arguments{
\item{opts}{file opts obtain with antaresRead::setSimulationPath}

\item{extname}{extention name for study duplicated}
}
\description{
Create An Area In An Antares Study
}
\examples{
\dontrun{

library(antaresRead)

# Set simulation path
opts = setSimulationPath(path = "PATH/TO/SIMULATION", simulation = "input")

# Create a new area
copyOputput(opts, "_adq")

}
}
