% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyOutput.R
\name{copyOutput}
\alias{copyOutput}
\title{Copy of the output files of an Antares study}
\usage{
copyOutput(opts, extname, mcYears = "all")
}
\arguments{
\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}

\item{extname}{Extension to be added to the name of the study, to be used as a name for the newly created folder.}

\item{mcYears}{mcYears to copy. Can be \code{"all"}.}
}
\description{
\verb{r antaresEditObject::badge_api_no()}

Copy of the output files of an Antares study.
}
\examples{
\dontrun{

library(antaresRead)

# Set simulation path
opts = setSimulationPath(path = "PATH/TO/SIMULATION", simulation = "input")

# Create a new area
copyOutput(opts, "_adq")

}

}
