% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importStudyAPI.R
\name{copyStudyWeb}
\alias{copyStudyWeb}
\title{Import physical study to Antares Web (managed study)}
\usage{
copyStudyWeb(
  opts = antaresRead::simOptions(),
  host,
  token,
  outputs = T,
  groups = NULL,
  suffix = "managedCopy"
)
}
\arguments{
\item{opts}{List of simulation parameters returned by the function \code{antaresRead::setSimulationPath}.
If id is not available, \code{antaresRead::searchStudy} will be used to find study.}

\item{host}{Host of AntaREST server API.}

\item{token}{API personnal access token.}

\item{outputs}{Logical. Determine if outputs are copied too.}

\item{groups}{Character. Add study to groups of Antares Web.}

\item{suffix}{Character. default is "managedCopy"
By default the new study will be : studyname_managedCopy}
}
\value{
New managed study ID
}
\description{
Copy study from an existing workspace into a managed study.
NOTE : The study must be present in a workspace (DRD, PPSE..) not just locally.
}
