% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarioBuilder.R
\name{create_scb_referential_series_type}
\alias{create_scb_referential_series_type}
\title{Create the correspondence data frame between the symbol and the type in scenario builder}
\usage{
create_scb_referential_series_type()
}
\value{
a \code{data.frame}.
}
\description{
Create the correspondence data frame between the symbol and the type in scenario builder
}
