% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCluster.R
\name{createCluster}
\alias{createCluster}
\title{Create a thermal cluster}
\usage{
createCluster(area, cluster_name, ..., time_series = NULL,
  prepro_data = NULL, prepro_modulation = NULL, add_prefix = TRUE,
  overwrite = FALSE, opts = antaresRead::simOptions())
}
\arguments{
\item{area}{The area where to create the cluster.}

\item{cluster_name}{cluster name.}

\item{...}{Parameters to write in the Ini file.}

\item{time_series}{the "ready-made" 8760-hour time-series available for simulation purposes.}

\item{prepro_data}{Pre-process data.}

\item{prepro_modulation}{Pre-process modulation.}

\item{add_prefix}{If \code{TRUE}, cluster_name will be prefixed by area's name.}

\item{overwrite}{Logical, overwrite the cluster or not.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
An updated list containing various information about the simulation.
}
\description{
Create a thermal cluster
}
\examples{
\dontrun{
createCluster(area = "fr", cluster_name = "fr_gas",
              group = "other", `marginal-cost` = 50)
}
}
