% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createClusterBulk.R
\name{createClusterBulk}
\alias{createClusterBulk}
\title{Create serial thermal cluster}
\usage{
createClusterBulk(
  cluster_object,
  add_prefix = TRUE,
  area_zone = "fr",
  opts = antaresRead::simOptions()
)
}
\arguments{
\item{cluster_object}{\code{list} native components of cluster}

\item{add_prefix}{\code{logical} prefix cluster name with area name}

\item{area_zone}{\code{character} name of area to create cluster}

\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}}
}
\value{
An updated list containing various information about the simulation.

\code{list} containing meta information about the simulation
}
\description{
For each area, the thermal cluster data are generated
}
\examples{
\dontrun{

library(antaresRead)
library(antaresEditObject)

# /!\/!\/!\ use or create a study /!\/!\/!\

# data preparation for sutructures
ts <- matrix(rep(c(0, 8000), each = 24*364), 
             ncol = 2)

df_pd <- matrix(rep(c(1, 1, 1, 0), each = 24*365), 
                ncol = 4)

df_pm <- matrix(data = c(rep(1, times = 365 * 24 * 3), rep(0, times = 365 * 24 * 1)), 
                ncol = 4)

# data structure for an area
zone_test_1 <- list(
  `CCGT old 1`= list(parameter= list(
    name= "CCGT old 1",
    group = "Other",
    unitcount= 10L,
    nominalcapacity= 100,
    enabled= "true",
   `min-stable-power`= 80L,
    `min-up-time`= 20L,
    `min-down_time`= 30L),
    time_series = ts,
    prepro_data = df_pd,
    prepro_modulation = df_pm),
  
  `CCGT old 2`= list(parameter= list(
    name= "CCGT old 2",
    group = "Other"),
    time_series = ts,
    prepro_data = df_pd,
    prepro_modulation = df_pm)
)

# Create multiple areas with multiple clusters
list_areas <- antaresRead::getAreas()[1:5]

lapply(list_areas, createClusterBulk,
cluster_object = c(zone_test_1, zone_test_2),
add_prefix = TRUE, 
opts = opts_temp)


}

}
