% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPSP.R
\name{createPSP}
\alias{createPSP}
\title{Create a Pumped Storage Power plant (PSP)}
\usage{
createPSP(areasAndCapacities = NULL, namePumping = "Psp_In",
  nameTurbining = "Psp_Out", hurdleCost = 5e-04,
  timeStepBindConstraint = "weekly", efficiency = NULL, overwrite = FALSE,
  opts = antaresRead::simOptions())
}
\arguments{
\item{areasAndCapacities}{A data.frame with 2 columns installedCapacity and area}

\item{namePumping}{The name of the pumping area}

\item{nameTurbining}{The name of the turbining area}

\item{hurdleCost}{The cost of the PSP}

\item{timeStepBindConstraint}{Time step for the binding constraint : \code{daily} or \code{weekly}}

\item{efficiency}{The efficiency of the PSP}

\item{overwrite}{Overwrite the Pumped Storage Power plant if already exist.
This will overwrite the previous area and links.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
An updated list containing various information about the simulation.
}
\description{
Create a Pumped Storage Power plant (PSP)
}
\examples{
\dontrun{

library(antaresEditObject)
path<-pathToYourStudy
opts<-setSimulationPath(path, simulation = "input")
pspData<-data.frame(area=c("a", "b"), installedCapacity=c(800,900))

createPSP(pspData, efficiency = 0.8)

createPSP(pspData, efficiency = 0.66, overwrite = TRUE)
createPSP(pspData, efficiency = 0.98, timeStepBindConstraint = "daily")
getAreas()

}

}
