% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStudy.R
\name{createStudy}
\alias{createStudy}
\title{Create an empty Antares study}
\usage{
createStudy(path, study_name = "my_study", antares_version = "7.0.0")
}
\arguments{
\item{path}{Path where to create study, it should be an empty directory,
if it doesn't exist, it'll be created.}

\item{study_name}{Name of the study.}

\item{antares_version}{Antares number version.}
}
\value{
logical vector indicating success or failure
}
\description{
Create an empty Antares study
}
\examples{
\dontrun{

createStudy("path/to/simulation")

}
}
