% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSimulation.R
\name{define_api_extra_options}
\alias{define_api_extra_options}
\title{Output profile options for running a simulation in API mode}
\usage{
define_api_extra_options(
  launcher = NULL,
  solver_preset = NULL,
  solver_version = NULL,
  run_at = NULL
)
}
\arguments{
\item{launcher}{Name of the launcher.}

\item{solver_preset}{Name of the solver.}

\item{solver_version}{Version of the solver.}

\item{run_at}{When the simulation should be run. Format YYYY-MM-DD hh-mm-ss. Time provided in GMT.}
}
\value{
a named list
}
\description{
Output profile options for running a simulation in API mode
}
\examples{
define_api_extra_options(
  launcher="calin2opf",
  solver_version="9.4"
)
}
