% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{detect_pattern_in_binding_constraint}
\alias{detect_pattern_in_binding_constraint}
\title{Detect a pattern in a binding constraint coefficient}
\usage{
detect_pattern_in_binding_constraint(pattern, opts = antaresRead::simOptions())
}
\arguments{
\item{pattern}{The pattern to detect.}

\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}}
}
\value{
An updated list containing various information about the simulation.

the names of the binding constraints containing the pattern
}
\description{
Detect a pattern in a binding constraint coefficient
}
