% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateAdequacySettings.R
\name{dicoAdequacySettings}
\alias{dicoAdequacySettings}
\title{Correspondence between arguments of \code{updateAdequacySettings} and actual Antares parameters.}
\usage{
dicoAdequacySettings(arg)
}
\arguments{
\item{arg}{An argument from function \code{updateAdequacySettings}.}
}
\value{
The corresponding Antares general parameter.
}
\description{
Correspondence between arguments of \code{updateAdequacySettings} and actual Antares parameters.
}
