% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCompatibilitySettings.R
\name{dicoCompatibilitySettings}
\alias{dicoCompatibilitySettings}
\title{Correspondence between arguments of \code{updateCompatibilitySettings} and actual Antares parameters.}
\usage{
dicoCompatibilitySettings(arg)
}
\arguments{
\item{arg}{An argument from function \code{updateCompatibilitySettings}.}
}
\value{
The corresponding Antares general parameter.
}
\description{
Correspondence between arguments of \code{updateCompatibilitySettings} and actual Antares parameters.
}
\examples{
dicoCompatibilitySettings("hydro.pmax") # "year-by-year"
}
