% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateGeneralSettings.R
\name{dicoGeneralSettings}
\alias{dicoGeneralSettings}
\title{Correspondence between arguments of \code{updateGeneralSettings} and actual Antares parameters.}
\usage{
dicoGeneralSettings(arg)
}
\arguments{
\item{arg}{An argument from function \code{updateGeneralSettings}.}
}
\value{
The corresponding Antares general parameter.
}
\description{
Correspondence between arguments of \code{updateGeneralSettings} and actual Antares parameters.
}
\examples{
dicoGeneralSettings("year.by.year") # "year-by-year"
}
