% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateOptimizationSettings.R
\name{dicoOptimizationSettings}
\alias{dicoOptimizationSettings}
\title{Correspondence between args of updateOptimizationSettings and actual Antares parameters.}
\usage{
dicoOptimizationSettings(arg)
}
\arguments{
\item{arg}{An argument from function \code{updateOptimizationSettings}.}
}
\value{
The corresponding Antares general parameter.
}
\description{
Correspondence between args of updateOptimizationSettings and actual Antares parameters.
}
\examples{
dicoGeneralSettings("year.by.year") # "year-by-year"
}
