% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateOutputSettings.R
\name{dicoOutputSettings}
\alias{dicoOutputSettings}
\title{Correspondence between arguments of \code{updateOutputSettings} and actual Antares parameters.}
\usage{
dicoOutputSettings(arg)
}
\arguments{
\item{arg}{An argument from function \code{updateOutputSettings}.}
}
\value{
The corresponding Antares general parameter.
}
\description{
Correspondence between arguments of \code{updateOutputSettings} and actual Antares parameters.
}
\examples{
dicoOutputSettings("result.format") # "result-format"
}
