% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editArea.R
\name{.api_command_execute_edit_area}
\alias{.api_command_execute_edit_area}
\title{Edit area's parameters in API mode.}
\usage{
.api_command_execute_edit_area(name, new_values, type, opts)
}
\arguments{
\item{name}{Name of the area to edit.}

\item{new_values}{Values of the parameters to edit.}

\item{type}{Type of edition.}

\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}}
}
\value{
An updated list containing various information about the simulation.
}
\description{
Edit area's parameters in API mode.
}
