% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editArea.R
\name{.api_command_execute_edit_area_ui}
\alias{.api_command_execute_edit_area_ui}
\title{Edit area's ui in API mode.}
\usage{
.api_command_execute_edit_area_ui(name, color, localization, opts)
}
\arguments{
\item{name}{Name of the area as a character, without punctuation except - and _.}

\item{color}{Color of the node}

\item{localization}{Localization on the map}

\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}}
}
\description{
Edit area's ui in API mode.
}
