% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTimeStepFromHourly.R
\name{.dailyToWeekly}
\alias{.dailyToWeekly}
\title{Computation function for rebuild mc-ind weekly from daily data.}
\usage{
.dailyToWeekly(dailydata, opts = simOptions(), type)
}
\arguments{
\item{dailydata}{antaresData for daily timestep}

\item{opts}{study opts}

\item{type}{type of data (areas, links, clusters, clustersRes)}
}
\description{
Computation function for rebuild mc-ind weekly from daily data.
}
\keyword{internal}
