% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.format_ini_rhs}
\alias{.format_ini_rhs}
\title{Format a value to a suitable format to rhs in an .ini file.}
\usage{
.format_ini_rhs(value)
}
\arguments{
\item{value}{The value to format.}
}
\value{
the formatted value
}
\description{
Format a value to a suitable format to rhs in an .ini file.
}
