% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlist.R
\name{.format_playlist_weights}
\alias{.format_playlist_weights}
\title{Generate playlist_year_weight section in the appropriate format.}
\usage{
.format_playlist_weights(weights, api_mode)
}
\arguments{
\item{weights}{data.table, 2 columns : mcYears and weights. Only with after antares V8}

\item{api_mode}{Boolean to identify an api study}
}
\value{
The playlist_year_weight section formatted.
}
\description{
Generate playlist_year_weight section in the appropriate format.
}
