% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createArea.R
\name{.format_ui_data_by_mode}
\alias{.format_ui_data_by_mode}
\title{Generate the data for the ui part by study mode.}
\usage{
.format_ui_data_by_mode(name, localization, color, api_mode)
}
\arguments{
\item{name}{Name of the area as a character, without punctuation except - and _.}

\item{localization}{Localization on the map}

\item{color}{Color of the node}

\item{api_mode}{Is study in API mode?}
}
\description{
Generate the data for the ui part by study mode.
}
