% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTimeStepFromHourly.R
\name{.hourlyToOther}
\alias{.hourlyToOther}
\title{Computation function for rebuild mc-ind daily, monthly and annual from hourly data.}
\usage{
.hourlyToOther(hourlydata, timeStep, type)
}
\arguments{
\item{hourlydata}{antaresData for hourly timestep}

\item{timeStep}{timestep of aggregation (daily, monthly and annual, NO weekly)}

\item{type}{type of data (areas, links, clusters, clustersRes)}
}
\description{
Computation function for rebuild mc-ind daily, monthly and annual from hourly data.
}
\keyword{internal}
