% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createArea.R
\name{.initializeLinksArea}
\alias{.initializeLinksArea}
\title{Initialize links data for a new area. For disk mode only.}
\usage{
.initializeLinksArea(name, overwrite, opts)
}
\arguments{
\item{name}{Name of the area as a character, without punctuation except - and _.}

\item{overwrite}{Overwrite the area if already exists.}

\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}}
}
\value{
An updated list containing various information about the simulation.
}
\description{
Initialize links data for a new area. For disk mode only.
}
