% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createArea.R
\name{.initializeThermalArea}
\alias{.initializeThermalArea}
\title{Initialize thermal data for a new area. For disk mode only.}
\usage{
.initializeThermalArea(name, overwrite, economic_options, opts)
}
\arguments{
\item{name}{Name of the area as a character, without punctuation except - and _.}

\item{overwrite}{Overwrite the area if already exists.}

\item{economic_options}{Economic options.}

\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}}
}
\value{
An updated list containing various information about the simulation.
}
\description{
Initialize thermal data for a new area. For disk mode only.
}
