% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStudy.R
\name{.is_version_9}
\alias{.is_version_9}
\title{function that ensures the transition to 9}
\usage{
.is_version_9(version)
}
\arguments{
\item{version}{\code{character} (eg, "8.2.0" or "9.0")}
}
\value{
\code{logical} if version >=9
}
\description{
basic tests on the version's writing format
}
\keyword{internal}
