% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createArea.R
\name{.split_nodalOptimization_by_target}
\alias{.split_nodalOptimization_by_target}
\title{Split list nodalOptimization by target file.}
\usage{
.split_nodalOptimization_by_target(nodalOptimization)
}
\arguments{
\item{nodalOptimization}{Nodal optimization parameters, see \code{\link[=nodalOptimizationOptions]{nodalOptimizationOptions()}}}
}
\description{
Split list nodalOptimization by target file.
}
