% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editArea.R
\name{editArea}
\alias{editArea}
\title{Edit An Existing Area In An Antares Study}
\usage{
editArea(
  name,
  color = NULL,
  localization = NULL,
  nodalOptimization = NULL,
  filtering = NULL,
  opts = antaresRead::simOptions()
)
}
\arguments{
\item{name}{Name of the area as a character, without punctuation except - and _.}

\item{color}{Color of the node}

\item{localization}{Localization on the map}

\item{nodalOptimization}{Nodal optimization parameters, see \code{\link{nodalOptimizationOptions}}.}

\item{filtering}{Filtering parameters, see \code{\link{filteringOptions}}.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
An updated list containing various information about the simulation.
}
\description{
Edit An Existing Area In An Antares Study
}
