% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeHydroValues.R
\name{fill_empty_hydro_ini_file}
\alias{fill_empty_hydro_ini_file}
\title{Write default values in hydro.ini file if the section is empty}
\usage{
fill_empty_hydro_ini_file(area, opts = antaresRead::simOptions())
}
\arguments{
\item{area}{The area where to write the value, i.e. lhs in the section.}

\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}.}
}
\description{
For a given area, if the data is empty, pick value from default values for \verb{use heuristic}, \verb{follow load} and \code{reservoir} sections.
}
