% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeHydroValues.R
\name{fill_empty_hydro_ts_file}
\alias{fill_empty_hydro_ts_file}
\title{Write default input time series if file is empty.}
\usage{
fill_empty_hydro_ts_file(area, opts = antaresRead::simOptions())
}
\arguments{
\item{area}{The area where to write the input time series.}

\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}.}
}
\description{
Write default input time series if file is empty.
}
