% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createArea.R
\name{filteringOptions}
\alias{filteringOptions}
\title{Output profile options for creating an area}
\usage{
filteringOptions(
  filter_synthesis = c("hourly", "daily", "weekly", "monthly", "annual"),
  filter_year_by_year = c("hourly", "daily", "weekly", "monthly", "annual")
)
}
\arguments{
\item{filter_synthesis}{Character, vector of time steps used in the output synthesis, among \code{hourly}, \code{daily}, \code{weekly}, \code{monthly}, and \code{annual}}

\item{filter_year_by_year}{Character, vector of time steps used in the output year-by-year, among \code{hourly}, \code{daily}, \code{weekly}, \code{monthly}, and \code{annual}}
}
\value{
a named list
}
\description{
Output profile options for creating an area
}
\examples{
filteringOptions(
  filter_synthesis=c("hourly","daily"),
  filter_year_by_year=c("weekly","monthly")
)
}
