% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeHydroValues.R
\name{get_default_hydro_ini_values}
\alias{get_default_hydro_ini_values}
\title{Get default hydro.ini values}
\usage{
get_default_hydro_ini_values(opts)
}
\arguments{
\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}}
}
\description{
Get default hydro.ini values
}
