% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeHydroValues.R
\name{get_type_check_mingen_vs_hydrostorage}
\alias{get_type_check_mingen_vs_hydrostorage}
\title{Get the type of control to execute using the 3 necessary booleans.}
\usage{
get_type_check_mingen_vs_hydrostorage(hydro_params)
}
\arguments{
\item{hydro_params}{a list of 3 booleans to compute the type of control to make.}
}
\value{
a list containing the type of control to execute and the variables to use for aggregation.
}
\description{
Get the type of control to execute using the 3 necessary booleans.
}
