% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeHydroValues.R
\name{get_type_check_mingen_vs_hydrostorage_to_trigger}
\alias{get_type_check_mingen_vs_hydrostorage_to_trigger}
\title{Get the type of control to execute between mingen data and hydro storage data}
\usage{
get_type_check_mingen_vs_hydrostorage_to_trigger(
  area,
  opts = antaresRead::simOptions()
)
}
\arguments{
\item{area}{The area where the type of control must be computed.}

\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}.}
}
\value{
a character containing the type of control to execute.
}
\description{
Compute the type of control to make between :
\itemize{
\item{\verb{input/hydro/series/<area>/mingen.txt}}
\item{\verb{input/hydro/series/<area>/mod.txt}}
}
This control is implemented in Antares too.
}
