% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{getAPIcommands}
\alias{getAPIcommands}
\title{Get API commands generated}
\usage{
getAPIcommands(opts = antaresRead::simOptions())
}
\arguments{
\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}}
}
\value{
a list of API commands
}
\description{
Get API commands generated
}
