% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{getJobLogs}
\alias{getJobLogs}
\title{Retrieve job log from API}
\usage{
getJobLogs(job_id, opts = antaresRead::simOptions())
}
\arguments{
\item{job_id}{The job identifier.}

\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}}
}
\value{
Logs as character string.
}
\description{
Retrieve job log from API
}
\examples{
\dontrun{

getJobLogs("37afff54-3a8b-40fa-8841-e1dae8854f5f")

}
}
