% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlist.R
\name{getPlaylist}
\alias{getPlaylist}
\title{Get the playlist of an Antares study}
\usage{
getPlaylist(opts = antaresRead::simOptions())
}
\arguments{
\item{opts}{list of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
Returns a vector of the identifier of the simulated MC year
}
\description{
\code{getPlaylist} gives the identifier of the MC years which
will be simulated in the Antares study, taking into account the potential use of a 
playlist which can skip some MC years
}
