% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCluster.R
\name{list_polluants_values}
\alias{list_polluants_values}
\title{Output polluants list for thermal clusters}
\usage{
list_polluants_values(multi_values = NULL)
}
\arguments{
\item{multi_values}{put values to init list values, default as \code{NULL}}
}
\value{
a named list
}
\description{
Output polluants list for thermal clusters
}
\examples{
list_polluants_values()
}
