% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createArea.R
\name{nodalOptimizationOptions}
\alias{nodalOptimizationOptions}
\title{Nodal optimization parameters for creating an area}
\usage{
nodalOptimizationOptions(
  non_dispatchable_power = TRUE,
  dispatchable_hydro_power = TRUE,
  other_dispatchable_power = TRUE,
  spread_unsupplied_energy_cost = 0,
  spread_spilled_energy_cost = 0,
  average_unsupplied_energy_cost = 0,
  average_spilled_energy_cost = 0
)
}
\arguments{
\item{non_dispatchable_power}{logical, default to FALSE}

\item{dispatchable_hydro_power}{logical, default to FALSE}

\item{other_dispatchable_power}{logical, default to FALSE}

\item{spread_unsupplied_energy_cost}{numeric, default to 0}

\item{spread_spilled_energy_cost}{numeric, default to 0}

\item{average_unsupplied_energy_cost}{numeric, default to 0}

\item{average_spilled_energy_cost}{numeric, default to 0}
}
\value{
a named list
}
\description{
Nodal optimization parameters for creating an area
}
\examples{
nodalOptimizationOptions()
}
