% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createClusterST.R
\name{normalize_ts_list}
\alias{normalize_ts_list}
\title{Normalize a named list of inputs to 8760×N matrices}
\usage{
normalize_ts_list(in_vals, allow_multi = FALSE)
}
\arguments{
\item{in_vals}{}

\item{allow_multi}{}
}
\description{
Normalize a named list of inputs to 8760×N matrices
}
