% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlist.R
\name{playlist}
\alias{playlist}
\alias{getPlaylist}
\alias{setPlaylist}
\title{Get the playlist of an Antares study}
\usage{
getPlaylist(opts = antaresRead::simOptions())

setPlaylist(playlist, weights = NULL, opts = antaresRead::simOptions())
}
\arguments{
\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}}

\item{playlist}{vector of MC years identifier to be simulated can be a list (V8 compatibility) but not recommended}

\item{weights}{data.table, 2 columns : mcYears and weights. Only with after antares V8}
}
\value{
\itemize{
\item \code{getPlaylist} returns a vector of the identifier of the simulated MC year
}

\itemize{
\item \code{setPlaylist} does not return anything. It is  used to modify the input of an Antares study.
}
}
\description{
\code{getPlaylist} gives the identifier of the MC years which
will be simulated in the Antares study, taking into account the potential use of a
playlist which can skip some MC years

\code{setPlaylist} is a function which modifies the input file of an ANTARES
study and set the playlist in order to simulate only the MC years given in input
}
