% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepSimulationWV.R
\name{prepSimulationWV}
\alias{prepSimulationWV}
\title{Prepare inputs to run a simulation for calculate water values}
\usage{
prepSimulationWV(area, fictive_area = paste0("WaterValue_", area),
  opts = antaresRead::simOptions())
}
\arguments{
\item{area}{The area studied}

\item{fictive_area}{Name of the fictive area to create}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
nothing yet
}
\description{
Prepare inputs to run a simulation for calculate water values
}
\examples{
\dontrun{
prepSimulationWV("fr")
}
}
