% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLink.R
\name{propertiesLink}
\alias{propertiesLink}
\title{Properties for creating a link}
\usage{
propertiesLink(hurdles_cost = FALSE, transmission_capacities = "enabled",
  display_comments = TRUE, filter_synthesis = c("hourly", "daily", "weekly",
  "monthly", "annual"), filter_year_by_year = c("hourly", "daily", "weekly",
  "monthly", "annual"))
}
\arguments{
\item{hurdles_cost}{Logical, which is used to state whether (linear)
transmission fees should be taken into account or not in economy and adequacy simulations}

\item{transmission_capacities}{Character, one of \code{enabled}, \code{infinite} or \code{ignore}, which is used to state whether 
the capacities to consider are those indicated in 8760-hour arrays or 
if zero or infinite values should be used instead (actual values / set to zero / set to infinite)}

\item{display_comments}{Logical}

\item{filter_synthesis}{Output synthesis}

\item{filter_year_by_year}{Output year-by-year}
}
\value{
A named list
}
\description{
Properties for creating a link
}
\examples{
\dontrun{
propertiesLink()
}
}
