% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLink.R
\name{propertiesLinkOptions}
\alias{propertiesLinkOptions}
\title{Properties for creating a link}
\usage{
propertiesLinkOptions(
  hurdles_cost = FALSE,
  transmission_capacities = "enabled",
  asset_type = "ac",
  display_comments = TRUE,
  filter_synthesis = c("hourly", "daily", "weekly", "monthly", "annual"),
  filter_year_by_year = c("hourly", "daily", "weekly", "monthly", "annual"),
  use_phase_shifter = FALSE,
  loop_flow = FALSE,
  colorr = 112,
  colorb = 112,
  colorg = 112,
  link_width = 1,
  link_style = "plain"
)
}
\arguments{
\item{hurdles_cost}{Logical, which is used to state whether (linear)
transmission fees should be taken into account or not in economy and adequacy simulations}

\item{transmission_capacities}{Character, one of \code{enabled}, \code{ignore} or \code{infinite}, which is used to state whether
the capacities to consider are those indicated in 8760-hour arrays or
if zero or infinite values should be used instead (actual values / set to zero / set to infinite)}

\item{asset_type}{Character, one of \code{ac}, \code{dc}, \code{gas}, \code{virt} or \code{other}. Used to
state whether the link is either an AC component (subject to Kirchhoff’s laws), a DC component,
or another type of asset.}

\item{display_comments}{Logical, display comments or not.}

\item{filter_synthesis}{Character, vector of time steps used in the output synthesis, among \code{hourly}, \code{daily}, \code{weekly}, \code{monthly}, and \code{annual}}

\item{filter_year_by_year}{Character, vector of time steps used in the output year-by-year, among \code{hourly}, \code{daily}, \code{weekly}, \code{monthly}, and \code{annual}}

\item{use_phase_shifter}{Logical.}

\item{loop_flow}{Logical.}

\item{colorr}{Integer, color of the line.}

\item{colorb}{Integer, color of the line.}

\item{colorg}{Integer, color of the line.}

\item{link_width}{Numeric, width of the line.}

\item{link_style}{Character, style of the line.}
}
\value{
A named list that can be used in \code{\link[=createLink]{createLink()}}.
}
\description{
Properties for creating a link
}
\examples{
\dontrun{
propertiesLinkOptions(
  hurdles_cost = TRUE,
  filter_synthesis=c("hourly","daily"),
  filter_year_by_year=c("weekly","monthly")
)
}
}
