% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIniFile.R
\name{readIniFile}
\alias{readIniFile}
\title{Read a INI file}
\usage{
readIniFile(file, stringsAsFactors = FALSE)
}
\arguments{
\item{file}{file path.}

\item{stringsAsFactors}{logical: should character vectors be converted to factors?}
}
\value{
A list with an element for each section of the .ini file.
}
\description{
Read a INI file
}
