% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeArea.R
\name{removeArea}
\alias{removeArea}
\title{Remove An Area From inputs}
\usage{
removeArea(name, opts = antaresRead::simOptions())
}
\arguments{
\item{name}{An area name}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
An upddated list containing various information about the simulation.
}
\description{
Remove An Area From inputs
}
\examples{
\dontrun{
removeArea("fictive_area")
}
}
