% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeBindingConstraint.R
\name{removeBindingConstraint}
\alias{removeBindingConstraint}
\title{Remove a Binding Constraint}
\usage{
removeBindingConstraint(name, opts = antaresRead::simOptions())
}
\arguments{
\item{name}{Name(s) of the binding constraint(s) to remove.}

\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}}
}
\value{
An updated list containing various information about the simulation.
}
\description{
\verb{r antaresEditObject::badge_api_ok()}

Remove a binding constraint in an Antares study.
}
\examples{
\dontrun{
removeBindingConstraint("mybindingconstraint")
}
}
\seealso{
\code{\link[=createBindingConstraint]{createBindingConstraint()}} to create new binding constraints, \code{\link[=editBindingConstraint]{editBindingConstraint()}} to edit existing binding constraints.
}
