% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/district.R
\name{removeDistrict}
\alias{removeDistrict}
\title{Remove a district}
\usage{
removeDistrict(name, opts = simOptions())
}
\arguments{
\item{name}{District's name.}

\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}}
}
\value{
An updated list containing various information about the simulation.
}
\description{
Remove a district
}
\examples{
\dontrun{
removeDistrict(
  name = "mydistrict",
  opts = simOptions()
)
}
}
