% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeLink.R
\name{removeLink}
\alias{removeLink}
\title{Remove a link between two areas}
\usage{
removeLink(from, to, opts = antaresRead::simOptions())
}
\arguments{
\item{from}{The first area from which to create a link}

\item{to}{The second one}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
An updated list containing various information about the simulation.
}
\description{
Remove a link between two areas
}
\examples{
\dontrun{
createLink(from = "myarea", to  = "myarea2")
removeLink(from = "myarea", to  = "myarea2")
}
}
