% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeOutput.R
\name{removeOutput}
\alias{removeOutput}
\title{Delete an Antares Simulation}
\usage{
removeOutput(name, opts)
}
\arguments{
\item{name}{Name of the simulation}

\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}}
}
\description{
\ifelse{html}{\figure{badge_api_ok.svg}{options: alt='Antares API OK'}}{Antares API: \strong{OK}}

Delete an Antares Simulation
}
