% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeHydroValues.R
\name{replicate_missing_ts}
\alias{replicate_missing_ts}
\title{Replicate a data.table as many times as needed to get the same number of time series between 2 data.tables.}
\usage{
replicate_missing_ts(xts, yts)
}
\arguments{
\item{x}{a data.table of time series type to replicate if necessary.}

\item{y}{a data.table of time series type to use as reference to match its number of time series.}
}
\value{
the data.table x replicated to match the number of time series of y.
}
\description{
Replicate a data.table as many times as needed to get the same number of time series between 2 data.tables.
}
