% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{setAPImode}
\alias{setAPImode}
\title{Set API mode}
\usage{
setAPImode(mode = c("async", "sync"), opts = antaresRead::simOptions())
}
\arguments{
\item{mode}{The mode you want to use.}

\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}}
}
\value{
An updated list containing various information about the simulation.
}
\description{
Two modes are available when using the API:
\itemize{
\item \strong{async}: record all API calls, but nothing is sent to the server
\item \strong{sync}: send query to the API each time a function is used
}
}
\examples{
\dontrun{
# See vignette for complete documentation
vignette("api-variant-management")

# Usage :
setAPImode("sync")

}
}
