% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSolverPath.R
\name{setSolverPath}
\alias{setSolverPath}
\title{Set path to Antares Solver}
\usage{
setSolverPath(path)
}
\arguments{
\item{path}{(optional) Path to the solver (e.g. 'antares-6.0-solver.exe' in \\bin directory where Antares is installed). 
If missing, a window opens and lets the user choose the directory of the simulation interactively.}
}
\description{
Set path to Antares Solver
}
\examples{

\dontrun{

setSolverPath(path = "C:/antares/bin/antares-6.0-solver.exe")

}

}
