% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createClusterST.R
\name{storage_values_default}
\alias{storage_values_default}
\title{Short Term Storage Property List}
\usage{
storage_values_default()
}
\value{
a named list
}
\description{
Short Term Storage Property List
}
\examples{
storage_values_default()
}
