% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCluster.R
\name{transform_list_to_json_for_createCluster}
\alias{transform_list_to_json_for_createCluster}
\title{Transform a user list to a json object to use in the endpoint of cluster creation}
\usage{
transform_list_to_json_for_createCluster(cluster_parameters, cluster_type)
}
\arguments{
\item{cluster_parameters}{a list containing the metadata of the cluster you want to create.}

\item{cluster_type}{the type of cluster you want to create. Each type has specific values.}
}
\value{
a json object
}
\description{
Transform a user list to a json object to use in the endpoint of cluster creation
}
