% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{update_generaldata_by_section}
\alias{update_generaldata_by_section}
\title{Update a specific section in generaldata.ini file}
\usage{
update_generaldata_by_section(opts, section, new_params)
}
\arguments{
\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}}

\item{section}{The section to update.}

\item{new_params}{The values to write in the section.}
}
\value{
An updated list containing various information about the simulation.
}
\description{
Update a specific section in generaldata.ini file
}
