% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateCompatibilitySettings.R
\name{updateCompatibilitySettings}
\alias{updateCompatibilitySettings}
\title{Update compatibility parameters of an Antares study}
\usage{
updateCompatibilitySettings(
  hydro.pmax = NULL,
  hydro.rule.curves = NULL,
  opts = antaresRead::simOptions()
)
}
\arguments{
\item{hydro.pmax}{daily or hourly}

\item{hydro.rule.curves}{single or scenarized}

\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}}
}
\value{
An updated list containing various information about the simulation.
}
\description{
\ifelse{html}{\figure{badge_api_ok.svg}{options: alt='Antares API OK'}}{Antares API: \strong{OK}}

Update compatibility parameters of an Antares study
}
\examples{
\dontrun{

updateCompatibilitySettings(
  hydro.pmax = "hourly", 
  hydro.rule.curves = "scenarized"
)

}
}
\seealso{
\href{https://antares-simulator.readthedocs.io/en/latest/user-guide/04-migration-guides/#new-optional-feature-and-compatibility-flag-for-scenarized-hydro-reservoir-levels}{Compatibility parameters}
}
