% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeAreaOutput.R
\name{write_area_output_values}
\alias{write_area_output_values}
\title{Write areas data}
\usage{
write_area_output_values(opts, data)
}
\arguments{
\item{opts}{opts simulation path}

\item{data}{obtain with readAntares}
}
\description{
This function write areas data for antares output
}
\examples{
\dontrun{


opts <- setSimulationPath("C:/Users/TitouanRobert/Desktop/Projet/RTE/antares/etude/MAF2025_S3_FBCOREMOD")
data <- readAntares(areas = "2_at00_hydro_open", mcYears = 1)
oot <- readAntares(mcYears = "all", area = "2_at00_hydro_open")
write_area_output_values(opts, oot)

}

}
