% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeLinkoutput.R
\name{write_link_output_values}
\alias{write_link_output_values}
\title{Write links data}
\usage{
write_link_output_values(opts, data)
}
\arguments{
\item{opts}{opts simulation path}

\item{data}{obtain with readAntares}
}
\description{
This function write links data for antares output
}
\examples{
\dontrun{
library(antaresRead)
library(data.table)
opts <- setSimulationPath("C:/Users/TitouanRobert/Desktop/Projet/RTE/antares/etude/MAF2025_S3_FBCOREMOD")
data <- readAntares(mcYears = 179, links = "all")
write_link_output_values(opts, data)

}

}
