% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeAntaresOutput.R
\name{writeOutputValues}
\alias{writeOutputValues}
\title{Write output value for antares}
\usage{
writeOutputValues(data, opts = NULL)
}
\arguments{
\item{data}{obtain with readAntares}

\item{opts}{opts simulation path}
}
\description{
This function write all ouput values for antares
}
\examples{
\dontrun{

library(antaresRead)
library(data.table)
opts <- setSimulationPath("my_study")
data <- readAntares(links = "all", areas = "all", clusters = "all")
writeOutputValues(data)


}

}
