% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupStudy.R
\name{backupStudy}
\alias{backupStudy}
\title{Create a backup with an Antares Study}
\usage{
backupStudy(backupfile, what = c("input", "study"),
  opts = antaresRead::simOptions())
}
\arguments{
\item{backupfile}{Name of the backup, without extension. If missing, 
either the name of the study or 'input' according argument \code{what}.}

\item{what}{Which folder to save, \code{input} for the input folder
or \code{study} for the whole study.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
The path of the backup
}
\description{
Save an Antares Study or only inputs in a .tar.gz file
}
\examples{
\dontrun{
backupStudy()
}
}
