% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTimeStampFromHourly.R
\name{computeTimeStampFromHourly}
\alias{computeTimeStampFromHourly}
\title{Compute daily, weekly, monthly and annual mc-ind from hourly data.}
\usage{
computeTimeStampFromHourly(opts, mcYears = "all", nbcl = 8, verbose = 1)
}
\arguments{
\item{opts}{opts simulation path}

\item{mcYears}{mcYears to compute}

\item{nbcl}{number of thread for parallel computing.}
}
\description{
Compute daily, weekly, monthly and annual mc-ind from hourly data.
}
\examples{
\dontrun{

library(antaresEditObject)
opts <- setSimulationPath("C:/Users/TitouanRobert/Desktop/Projet/RTE/antares/etude/new/BP19_FB18_2021_60mcAcc/output/20201015-0957eco-test3a")
computeTimeStampFromHourly(opts)

}

}
