% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyOutput.R
\name{copyOutput}
\alias{copyOutput}
\title{Copy an output studies.}
\usage{
copyOutput(opts, extname, mcYears = "all")
}
\arguments{
\item{opts}{file opts obtain with antaresRead::setSimulationPath}

\item{extname}{extention name for study duplicated}

\item{mcYears}{mcYears to copy. Can be all.}
}
\description{
Copy an output studies.
}
\examples{
\dontrun{

library(antaresRead)

# Set simulation path
opts = setSimulationPath(path = "PATH/TO/SIMULATION", simulation = "input")

# Create a new area
copyOutput(opts, "_adq")

}
}
