% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importStudyAPI.R
\name{copyStudyWeb}
\alias{copyStudyWeb}
\title{Import physical study to Antares Web (managed study)}
\usage{
copyStudyWeb(
  study_id,
  studyPath,
  opts = antaresRead::simOptions(),
  host,
  token,
  outputs = T,
  groups = NULL,
  suffix = "managedCopy"
)
}
\arguments{
\item{study_id}{Id of study if available.
If not, \code{antaresRead::searchStudy} will be used to find it.}

\item{studyPath}{Path of the study to search for.}

\item{opts}{List of simulation parameters returned by the function \code{antaresRead::setSimulationPath}.
Only used when both id and path is missing.}

\item{host}{Host of AntaREST server API.}

\item{token}{API personnal access token.}

\item{outputs}{Logical. Determine if outputs are copied too.}

\item{groups}{Character. Add study to groups of Antares Web.}

\item{suffix}{Character. default is "managedCopy"
By default the new study will be : studyname_managedCopy}
}
\value{
New managed study ID
}
\description{
Copy study from an existing workspace into a managed study.
NOTE : The study must be present in a workspace (DRD, PPSE..) not just locally.
}
