% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createArea.R
\name{createArea}
\alias{createArea}
\title{Create An Area In An Antares Study}
\usage{
createArea(name, color = grDevices::rgb(230, 108, 44, max = 255),
  localization = c(0, 0), nodalOptimization = nodalOptimizationOptions(),
  filtering = filteringOptions(), overwrite = FALSE,
  opts = antaresRead::simOptions())
}
\arguments{
\item{name}{Name of the area as a character, without punctuation except - and _.}

\item{color}{Color of the node}

\item{localization}{Localization on the map}

\item{nodalOptimization}{Nodal optimization parameters}

\item{filtering}{Filtering parameters}

\item{overwrite}{Overwrite the area if already exist}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
An updated list containing various information about the simulation.
}
\description{
Create An Area In An Antares Study
}
\examples{
\dontrun{

createArea("fictive_area")

}
}
