% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCluster.R
\name{createCluster}
\alias{createCluster}
\title{Create a thermal cluster}
\usage{
createCluster(area, cluster_name, ..., time_series = NULL,
  prepro_data = NULL, prepro_modulation = NULL, overwrite = FALSE,
  opts = antaresRead::simOptions())
}
\arguments{
\item{area}{The area where to create the cluster}

\item{cluster_name}{cluster name}

\item{...}{Parameters to write in the Ini file}

\item{time_series}{Cluster's time series}

\item{prepro_data}{Preprocess data}

\item{prepro_modulation}{Preprocess modulation}

\item{overwrite}{Logical, overwrite the cluster or not.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
Nothing
}
\description{
Create a thermal cluster
}
\examples{
\dontrun{
createCluster(area = "fr", cluster_name = "fr_gas",
              group = "other", `marginal-cost` = 50)
}
}
