% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createClusterST.R
\name{createClusterST}
\alias{createClusterST}
\title{Create a short-term storage cluster}
\usage{
createClusterST(
  area,
  cluster_name,
  group = "Other",
  ...,
  PMAX_injection = NULL,
  PMAX_withdrawal = NULL,
  inflows = NULL,
  lower_rule_curve = NULL,
  upper_rule_curve = NULL,
  add_prefix = TRUE,
  overwrite = FALSE,
  opts = antaresRead::simOptions()
)
}
\arguments{
\item{area}{The area where to create the cluster.}

\item{cluster_name}{Name for the cluster, it will prefixed by area name, unless you set \code{add_prefix = FALSE}.}

\item{group}{Group of the cluster, one of : "PSP_open", "PSP_closed", "Pondage", "Battery", "Other". It corresponds to the type of stockage.}

\item{...}{Parameters to write in the Ini file. Careful!
Some parameters must be set as \code{integers} to avoid warnings in Antares, for example,
to set \code{unitcount}, you'll have to use \code{unitcount = 1L}.}

\item{PMAX_injection}{modulation of charging capacity on an 8760-hour basis. The values are float between 0 and 1.}

\item{PMAX_withdrawal}{modulation of discharging capacity on an 8760-hour basis. The values are float between 0 and 1.}

\item{inflows}{imposed withdrawals from the stock for other uses, The values are integer.}

\item{lower_rule_curve}{This is the lower limit for filling the stock imposed each hour. The values are float between 0 and 1.}

\item{upper_rule_curve}{This is the upper limit for filling the stock imposed each hour. The values are float between 0 and 1.}

\item{add_prefix}{If \code{TRUE} (the default), \code{cluster_name} will be prefixed by area name.}

\item{overwrite}{Logical, overwrite the cluster or not.}

\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}}
}
\value{
An updated list containing various information about the simulation.
}
\description{
\ifelse{html}{\figure{badge_api_no.svg}{options: alt='Antares API NO'}}{Antares API: \strong{NO}}

Create a new ST-storage cluster for >= v8.6.0 Antares studies.
}
\note{
five files are written in output according to the input parameters
\itemize{
\item PMAX-injection.txt
\item PMAX-withdrawal.txt
\item inflows.txt
\item lower-rule-curve.txt
\item upper-rule-curve.txt
}
}
\examples{
\dontrun{

library(antaresRead)
library(antaresEditObject)

# Create a cluster :
createClusterST(
  area = "fr", 
  cluster_name = "my_cluster",
  group = "other", 
  unitcount = 1L, # or as.integer(1)
)
# by default, cluster name is prefixed 
# by the area name
levels(readClusterSTDesc()$cluster)
# > "fr_my_cluster"
}

}
\seealso{
\code{\link[=editClusterST]{editClusterST()}} to edit existing clusters, \code{\link[=removeClusterST]{removeClusterST()}} to remove clusters.
}
