% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStudy.R
\name{createStudy}
\alias{createStudy}
\alias{createStudyAPI}
\title{Create an empty Antares study}
\usage{
createStudy(path, study_name = "my_study", antares_version = "8.2.0")

createStudyAPI(
  host,
  token = NULL,
  study_name = "my_study",
  antares_version = "8.2.0",
  ...
)
}
\arguments{
\item{path}{Path where to create study, it should be an empty directory,
if it doesn't exist, it'll be created.}

\item{study_name}{Name of the study.}

\item{antares_version}{Antares number version.}

\item{host}{Host of AntaREST server API.}

\item{token}{API personnal access token.}

\item{...}{Other query parameters passed to POST request.}
}
\value{
Result of \code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}} or \code{\link[=setSimulationPathAPI]{setSimulationPathAPI()}} accordingly.
}
\description{
Create study on disk or with AntaREST server through the API.
}
\examples{
\dontrun{

createStudy("path/to/simulation")

}
}
