% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStudy.R
\name{deleteStudy}
\alias{deleteStudy}
\title{Delete a study}
\usage{
deleteStudy(opts = simOptions(), prompt_validation = FALSE)
}
\arguments{
\item{opts}{List. study options}

\item{prompt_validation}{\code{logical} to put validation message to delete study (default \code{FALSE})}
}
\description{
Delete a study
}
