% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editClusterST.R
\name{editClusterST}
\alias{editClusterST}
\title{Edit a short-term storage cluster}
\usage{
editClusterST(
  area,
  cluster_name,
  group = NULL,
  storage_parameters = NULL,
  PMAX_injection = NULL,
  PMAX_withdrawal = NULL,
  inflows = NULL,
  lower_rule_curve = NULL,
  upper_rule_curve = NULL,
  cost_injection = NULL,
  cost_withdrawal = NULL,
  cost_level = NULL,
  cost_variation_injection = NULL,
  cost_variation_withdrawal = NULL,
  add_prefix = TRUE,
  opts = antaresRead::simOptions()
)
}
\arguments{
\item{area}{The area where to create the cluster.}

\item{cluster_name}{Name for the cluster, it will prefixed by area name, unless you set \code{add_prefix = FALSE}.}

\item{group}{Group of the cluster, one of : \emph{{PSP_open, PSP_closed, Pondage, Battery, Other}}.
It corresponds to the type of stockage (\strong{dynamic name for Antares version >= 9.2}).}

\item{storage_parameters}{\code{list } Parameters to write in the Ini file (see \code{Note}).}

\item{PMAX_injection}{Modulation of charging capacity on an 8760-hour basis. \code{numeric} \{0;1\} (8760*1).}

\item{PMAX_withdrawal}{Modulation of discharging capacity on an 8760-hour basis. \code{numeric} \{0;1\} (8760*1).}

\item{inflows}{Algebraic deviation of the state of charge of the storage, which does not induce any power
generation or consumption on the system \code{numeric} \{<0;>0\} (8760*1).}

\item{lower_rule_curve}{This is the lower limit for filling the stock imposed each hour. \code{numeric} \{0;1\} (8760*1).}

\item{upper_rule_curve}{This is the upper limit for filling the stock imposed each hour. \code{numeric} \{0;1\} (8760*1).}

\item{cost_injection}{NULL}

\item{cost_withdrawal}{NULL}

\item{cost_level}{NULL}

\item{cost_variation_injection}{NULL}

\item{cost_variation_withdrawal}{NULL}

\item{add_prefix}{If \code{TRUE} (the default), \code{cluster_name} will be prefixed by area name.}

\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}}
}
\value{
An updated list containing various information about the simulation.
}
\description{
\ifelse{html}{\figure{badge_api_ok.svg}{options: alt='Antares API OK'}}{Antares API: \strong{OK}}

Edit parameters and time series of an existing \code{st-storage} cluster (Antares studies >= v8.6.0).
}
\note{
Put only properties or TS value you want to edit (see \code{examples} section).
}
\examples{
\dontrun{
# TODO
print(2+2)
}
}
\seealso{
\code{\link[=createClusterST]{createClusterST()}}, \code{\link[=removeClusterST]{removeClusterST()}}
}
