% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeHydroValues.R
\name{fill_empty_hydro_ini_file}
\alias{fill_empty_hydro_ini_file}
\title{Write default values in hydro.ini file if the section is empty.}
\usage{
fill_empty_hydro_ini_file(area, opts = antaresRead::simOptions())
}
\arguments{
\item{area}{The area where to write the value, i.e. lhs in the section.}

\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}.}
}
\description{
\ifelse{html}{\figure{badge_api_ok.svg}{options: alt='Antares API OK'}}{Antares API: \strong{OK}}

For a given area, if the data is empty, write TRUE for \verb{use heuristic} and \verb{follow load} sections and FALSE for \code{reservoir} section.
}
