% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createArea.R
\name{filteringOptions}
\alias{filteringOptions}
\title{Output profile options for creating an area}
\usage{
filteringOptions(filter_synthesis = c("hourly", "daily", "weekly", "monthly",
  "annual"), filter_year_by_year = c("hourly", "daily", "weekly", "monthly",
  "annual"))
}
\arguments{
\item{filter_synthesis}{Output synthesis}

\item{filter_year_by_year}{Output Year-by-year}
}
\value{
a named list
}
\description{
Output profile options for creating an area
}
\examples{
filteringOptions()
}
