% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{getJobs}
\alias{getJobs}
\title{Retrieve API jobs}
\usage{
getJobs(opts = antaresRead::simOptions())
}
\arguments{
\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}}
}
\value{
A \code{data.table} with information about jobs.
}
\description{
Retrieve API jobs
}
\examples{
\dontrun{

getJobs()

}
}
