% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importStudyAPI.R
\name{importZipStudyWeb}
\alias{importZipStudyWeb}
\title{Import a local study to Antares Web}
\usage{
importZipStudyWeb(
  host,
  token,
  zipfile_name,
  delete_zipfile = TRUE,
  folder_destination = NULL,
  opts = antaresRead::simOptions()
)
}
\arguments{
\item{host}{Host of AntaREST server API.}

\item{token}{API personnal access token.}

\item{zipfile_name}{Name of the zipfile of the study.}

\item{delete_zipfile}{Should the zipfile be deleted after upload.}

\item{folder_destination}{Folder of the study in Antares Web.}

\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}}
}
\value{
An updated list containing various information about the simulation.
}
\description{
Import a local study to Antares Web
}
