% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{mockSimulationAPI}
\alias{mockSimulationAPI}
\title{Mock API usage}
\usage{
mockSimulationAPI()
}
\value{
An updated list containing various information about the simulation.
}
\description{
Use this to generate command without an active API connection,
it allow to use function to edit a study to later on get API commands.
}
\examples{
\dontrun{
# Mock simulation API
mockSimulationAPI()
# Create an area
createArea("new area")
# Get commands
getVariantCommands()
}
}
