% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeArea.R
\name{removeArea}
\alias{removeArea}
\title{Remove an area from an Antares study}
\usage{
removeArea(name, opts = antaresRead::simOptions())
}
\arguments{
\item{name}{An area name.}

\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}}
}
\value{
An updated list containing various information about the simulation.
}
\description{
\verb{r antaresEditObject::badge_api_ok()}

Remove an area in an Antares study.
}
\examples{
\dontrun{
removeArea("fictive_area")
}
}
\seealso{
\code{\link[=createArea]{createArea()}}, \code{\link[=editArea]{editArea()}}
}
