% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeBindingConstraint.R
\name{removeBindingConstraint}
\alias{removeBindingConstraint}
\title{Remove a Binding Constraint}
\usage{
removeBindingConstraint(name, opts = antaresRead::simOptions())
}
\arguments{
\item{name}{Name(s) of the binding constraint(s) to remove.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
An upddated list containing various information about the simulation.
}
\description{
Remove a Binding Constraint
}
\examples{
\dontrun{
removeBindingConstraint("mybindingconstraint")
}
}
