% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeBindingConstraint.R
\name{removeBindingConstraint}
\alias{removeBindingConstraint}
\title{Remove a Binding Constraint}
\usage{
removeBindingConstraint(name, opts = antaresRead::simOptions())
}
\arguments{
\item{name}{Name(s) of the binding constraint(s) to remove.}

\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}}
}
\value{
An updated list containing various information about the simulation.
}
\description{
\ifelse{html}{\figure{badge_api_ok.svg}{options: alt='Antares API OK'}}{Antares API: \strong{OK}}

Remove a binding constraint in an Antares study.
}
\examples{
\dontrun{
removeBindingConstraint("mybindingconstraint")
}
}
