% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeLink.R
\name{removeLink}
\alias{removeLink}
\title{Remove a link between two areas}
\usage{
removeLink(from, to, opts = antaresRead::simOptions())
}
\arguments{
\item{from, to}{The two areas linked together.}

\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}}
}
\value{
An updated list containing various information about the simulation.
}
\description{
\verb{r antaresEditObject::badge_api_ok()}

Remove a link between two areas in an Antares study.
}
\examples{
\dontrun{
createLink(from = "myarea", to  = "myarea2")
removeLink(from = "myarea", to  = "myarea2")
}
}
