% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/playlist.R
\name{setPlaylist}
\alias{setPlaylist}
\title{Set the playlist of an Antares Study}
\usage{
setPlaylist(playlist, weights = NULL, opts = antaresRead::simOptions())
}
\arguments{
\item{playlist}{vector of MC years identifier to be simulated can be a list (V8 compatibility) but not recommended}

\item{weights}{data.table, 2 columns : mcYears and weights. Only with after antares V8}

\item{opts}{list of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
The function does not return anything. It is  used to modify the input of an 
Antares study
}
\description{
\code{set_playlist} is a function which modifies the input file of an ANTARES
study and set the playlist in order to simulate only the MC years given in input
}
