% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createClusterST.R
\name{storage_values_default}
\alias{storage_values_default}
\title{Output polluants list for thermal clusters}
\usage{
storage_values_default()
}
\value{
a named list
}
\description{
Output polluants list for thermal clusters
}
\examples{
storage_values_default()
}
