% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateInputSettings.R
\name{updateInputSettings}
\alias{updateInputSettings}
\title{Update input parameters of an Antares study}
\usage{
updateInputSettings(import, opts = antaresRead::simOptions())
}
\arguments{
\item{import}{Series to import.}

\item{opts}{List of simulation parameters returned by the function
\code{antaresRead::setSimulationPath}}
}
\value{
An updated list containing various information about the simulation.
}
\description{
Update input parameters of an Antares study
}
\examples{
\dontrun{

updateInputSettings(import = c("thermal"))
updateInputSettings(import = c("hydro", "thermal"))

}
}
