% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/API.R
\name{variant}
\alias{variant}
\alias{createVariant}
\alias{useVariant}
\title{Create a study's variant}
\usage{
createVariant(name, opts = antaresRead::simOptions())

useVariant(name, opts = antaresRead::simOptions())
}
\arguments{
\item{name}{Name for the variant to create or the name of an existent variant.}

\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}}
}
\value{
An updated list containing various information about the simulation.
}
\description{
\strong{API}: create a new variant for a given study or use a pre-existing one.
}
\examples{
\dontrun{
# See vignette for complete documentation
vignette("api-variant-management")
}
}
