% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeHydroValues.R
\name{writeIniHydro}
\alias{writeIniHydro}
\title{Edit hydro.ini values}
\usage{
writeIniHydro(area, params, mode = "other", opts = antaresRead::simOptions())
}
\arguments{
\item{area}{The area where to edit the values.}

\item{params}{The list data must have specific names and specific types :
\itemize{
\item{follow load}{ : logical or NULL}
\item{use heuristic}{ : logical or NULL}
\item{use water}{ : logical or NULL}
\item{hard bounds}{ : logical or NULL}
\item{use leeway}{ : logical or NULL}
\item{power to level}{ : logical or NULL}
\item{reservoir}{ : logical or NULL}
\item{inter-daily-breakdown}{ : numeric, integer or NULL}
\item{intra-daily-modulation}{ : numeric, integer or NULL}
\item{inter-monthly-breakdown}{ : numeric, integer or NULL}
\item{leeway low}{ : numeric, integer or NULL}
\item{leeway up}{ : numeric, integer or NULL}
\item{pumping efficiency}{ : numeric, integer or NULL}
\item{initialize reservoir date}{ : numeric, integer or NULL}
\item{reservoir capacity}{ : numeric, integer or NULL}
}}

\item{mode}{Execution mode. Useful when you create a new area or remove an existing area to avoid control on hydro data.}

\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}.}
}
\description{
\ifelse{html}{\figure{badge_api_ok.svg}{options: alt='Antares API OK'}}{Antares API: \strong{OK}}

For a given area, write its data in the hydro.ini file.
}
\examples{
\dontrun{
opts <- setSimulationPath(studypath, simulation = "input")
createArea("fictive_area") 
writeIniHydro(area = "fictive_area"
, params = list("leeway low" = 2.5, "leeway up" = 25))

}
}
