% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeInputTS.R
\name{writeInputTS}
\alias{writeInputTS}
\title{Write input time series}
\usage{
writeInputTS(
  data,
  type = c("load", "hydroROR", "hydroSTOR", "wind", "solar", "tsLink", "mingen"),
  area = NULL,
  link = NULL,
  overwrite = TRUE,
  opts = antaresRead::simOptions()
)
}
\arguments{
\item{data}{A 8760*N matrix of hourly time series, except when \code{type} is
\code{"hydroSTOR"}.
In this latter case \code{"hydroSTOR"} data must have either be 365 rows
(Antares v7) or 12 rows (v6 and earlier).

In the \code{"mingen"} case, \code{"antaresVersion"} should be >= 860.

\code{"hydroSTOR"} and \code{"mingen"} number of columns must be the same :

If the \code{"hydroSTOR"} (\code{"mingen"}) data consists of only one column,
then there are no constraints on the number of \code{"mingen"} (\code{"hydroSTOR"}) columns.

If the \code{"hydroSTOR"} (\code{"mingen"}) data is composed of more than a one column,
then \code{"mingen"} (\code{"hydroSTOR"}) data must be composed of either one column,
or the same number as \code{"hydroSTOR"} (\code{"mingen"}) data.}

\item{type}{Serie to write: \code{"load"}, \code{"hydroROR"}, \code{"hydroSTOR"},
\code{"wind"}, \code{"solar"}, \code{"tsLink"} or \code{"mingen"}.}

\item{area}{The area where to write the input time series.}

\item{link}{Link for which writing transmission capacities time series,
must like \code{"area01\%area02"} or \code{c("area01", "area02")}.}

\item{overwrite}{Logical. Overwrite the values if a file already exists.}

\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}}
}
\value{
An updated list containing various information about the simulation.
}
\description{
\ifelse{html}{\figure{badge_api_ok.svg}{options: alt='Antares API OK'}}{Antares API: \strong{OK}}

This function writes input time series in an Antares project.
}
\section{Warning}{


You cannot use \code{area} and \code{link} arguments at the same time.
}

\examples{
\dontrun{

# Write solar time series
writeInputTS(
  area = "fictive_area",
  type = "solar",
  data = matrix(rep(4, 8760*2), nrow = 8760)
)

}
}
