% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeAntaresOutput.R
\name{writeOutputValues}
\alias{writeOutputValues}
\title{Write output value for Antares}
\usage{
writeOutputValues(data, opts = NULL)
}
\arguments{
\item{data}{obtain with readAntares}

\item{opts}{List of simulation parameters returned by the function
\code{\link[antaresRead:setSimulationPath]{antaresRead::setSimulationPath()}}}
}
\description{
\verb{r antaresEditObject::badge_api_no()}

This function write all output values for an Antares study.
}
\examples{
\dontrun{

library(antaresRead)
library(data.table)
opts <- setSimulationPath("my_study")
data <- readAntares(links = "all", areas = "all", clusters = "all")
writeOutputValues(data)

}

}
